Split Layer stories
