+++ src/js/components/Layer/stories/ScrollBody.js
@@ -1,0 +1,104 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { FormClose } from 'grommet-icons';
+
+import { Box, Button, Grommet, Layer, Text } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const ScrollBodyLayer = () => (
+  <Grommet theme={grommet}>
+    <Layer full="vertical" position="right">
+      <Box fill style={{ minWidth: '378px' }}>
+        <Box
+          direction="row"
+          align="center"
+          as="header"
+          elevation="small"
+          justify="between"
+        >
+          <Text margin={{ left: 'small' }}>Header</Text>
+          <Button icon={<FormClose />} />
+        </Box>
+        <Box flex overflow="auto" pad="xsmall">
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+          <span>body</span>
+        </Box>
+        <Box
+          as="footer"
+          border={{ side: 'top' }}
+          pad="small"
+          justify="end"
+          direction="row"
+          align="center"
+        >
+          <Button primary label="Save" />
+        </Box>
+      </Box>
+    </Layer>
+  </Grommet>
+);
+storiesOf('Layer', module).add('Fixed Header, Scroll Body', () => (
+  <ScrollBodyLayer />
+));
